requirements.txt
Pillow==10.1.0
torch==2.1.2
torchvision==0.16.2
transformers==4.40.0
sentencepiece==0.1.99
accelerate==0.30.1
bitsandbytes==0.43.1
opencv-python==4.10.0.82
cog==0.9.12
numpy==1.26.4